# evcar-news/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/markets": {
      "get": {
        "operationId": "Markets_list",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Overrides Accept-Language. Default ar.",
            "schema": {
              "enum": [
                "ar",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "market",
            "required": false,
            "in": "query",
            "description": "Market code (e.g. EG). Overrides X-Market.",
            "schema": {}
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete list (not paginated): meta is page 1 of 1 with pageSize = total.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PublicMarketDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Enabled markets with their currency (names in the request language)",
        "tags": [
          "markets"
        ]
      }
    },
    "/api/v1/markets/{code}": {
      "get": {
        "operationId": "Markets_get",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Overrides Accept-Language. Default ar.",
            "schema": {
              "enum": [
                "ar",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "market",
            "required": false,
            "in": "query",
            "description": "Market code (e.g. EG). Overrides X-Market.",
            "schema": {}
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PublicMarketDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "markets"
        ]
      }
    },
    "/api/v1/admin/markets": {
      "get": {
        "operationId": "AdminMarkets_list",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete list (not paginated): meta is page 1 of 1 with pageSize = total.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminMarketDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "All markets including disabled ones",
        "tags": [
          "admin-markets"
        ]
      },
      "post": {
        "operationId": "AdminMarkets_create",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminMarketDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Adds a country (starts disabled unless enabled=true)",
        "tags": [
          "admin-markets"
        ]
      }
    },
    "/api/v1/admin/markets/{code}": {
      "get": {
        "operationId": "AdminMarkets_get",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminMarketDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "One market with its usage counts",
        "tags": [
          "admin-markets"
        ]
      },
      "patch": {
        "operationId": "AdminMarkets_update",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminMarketDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Edits / enables / disables a market (the default cannot be disabled)",
        "tags": [
          "admin-markets"
        ]
      },
      "delete": {
        "operationId": "AdminMarkets_remove",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deletes an unused market (otherwise 409 MARKET_IN_USE: disable it)",
        "tags": [
          "admin-markets"
        ]
      }
    },
    "/api/v1/admin/currencies": {
      "get": {
        "operationId": "AdminCurrencies_list",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete list (not paginated): meta is page 1 of 1 with pageSize = total.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CurrencyDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin-markets"
        ]
      },
      "post": {
        "operationId": "AdminCurrencies_create",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CurrencyDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin-markets"
        ]
      }
    },
    "/api/v1/admin/currencies/{code}": {
      "patch": {
        "operationId": "AdminCurrencies_update",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CurrencyDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin-markets"
        ]
      },
      "delete": {
        "operationId": "AdminCurrencies_remove",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deletes an unused currency (otherwise 409 CURRENCY_IN_USE)",
        "tags": [
          "admin-markets"
        ]
      }
    },
...
  },
  "info": {
    "title": "EV Car News API",
    "description": "REST API of EV Car News (evcar.news). Conventions:\n- Prefix /api/v1; admin routes under /api/v1/admin (permission-guarded), personal routes under /api/v1/me.\n- Language: ?lang=ar|en overrides Accept-Language (default ar). Market: ?market=EG overrides X-Market.\n- Single resource: {\"data\": {...}}; lists: {\"data\": [...], \"meta\": {page, pageSize, total, totalPages}}.\n- Errors: {\"error\": {\"code\", \"message\" (localized), \"details\"?, \"requestId\"}}; validation → 422 VALIDATION_FAILED.\n- Money: {\"amount\": \"decimal string\", \"currency\": \"EGP\"}; missing values are null (never 0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "PublicMarketDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "EG"
          },
          "name": {
            "type": "string",
            "description": "Name in the request language.",
            "example": "مصر"
          },
          "nameAr": {
            "type": "string",
            "example": "مصر"
          },
          "nameEn": {
            "type": "string",
            "example": "Egypt"
          },
          "currency": {
            "$ref": "#/components/schemas/PublicCurrencyDto"
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Cairo"
          },
          "defaultLanguage": {
            "type": "string",
            "enum": [
              "ar",
              "en"
            ],
            "description": "Suggested language only — language and market are independent."
          },
          "unitSystem": {
            "type": "string",
            "enum": [
              "metric",
              "imperial"
            ]
          },
          "driveSide": {
            "type": "string",
            "enum": [
              "lhd",
              "rhd"
            ]
          },
          "isDefault": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "name",
          "nameAr",
          "nameEn",
          "currency",
          "timezone",
          "defaultLanguage",
          "unitSystem",
          "driveSide",
          "isDefault",
          "sortOrder"
        ]
      },
      "CurrencyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "EGP"
          },
          "nameAr": {
            "type": "string",
            "example": "جنيه مصري"
          },
          "nameEn": {
            "type": "string",
            "example": "Egyptian Pound"
          },
          "symbolAr": {
            "type": "string",
            "nullable": true,
            "example": "ج.م"
          },
          "symbolEn": {
            "type": "string",
            "nullable": true,
            "example": "E£"
          },
          "decimals": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "code",
          "nameAr",
          "nameEn",
          "symbolAr",
          "symbolEn",
          "decimals"
        ]
      },
...
    }
  }
}